Improve skeletons

+++ src/drom_lib/commandPackage.ml
@@ -15,24 +15,28 @@ open EzFile.OP
 let cmd_name = "package"
 
 let remove_file hashes file =
-  try
-    if Sys.is_directory file then
-      Unix.rmdir file
-    else begin
-      Sys.remove file;
-      Hashes.remove hashes file
-    end
-  with exn ->
-    Printf.eprintf "remove %s failed with %s\n%!"
-      file (Printexc.to_string exn)
+  if Sys.file_exists file then
+    try
+      if Sys.is_directory file then
+        Unix.rmdir file
+      else begin
+        Sys.remove file;
+        Hashes.remove hashes file
+      end
+    with exn ->
+      Printf.eprintf "remove %s failed with %s\n%!"
+        file (Printexc.to_string exn)
+  else
+    Hashes.remove hashes file
 
 let remove_dir hashes dir =
-  EzFile.make_select EzFile.iter_dir dir
-    ~deep:true ~dft:`After
-    ~f:(fun path ->
-        let file = Filename.concat dir path in
-        remove_file hashes file;
-      )
+  if Sys.file_exists dir then
+    EzFile.make_select EzFile.iter_dir dir
+      ~deep:true ~dft:`After
+      ~f:(fun path ->
+          let file = Filename.concat dir path in
+          remove_file hashes file;
+        )
 
 let rename_dir hashes src dst =
   EzFile.make_dir ~p:true dst ;
@@ -61,7 +65,8 @@ let rename_dir hashes src dst =
 let remove_package hashes package =
   remove_dir hashes package.dir;
   let file = package.name ^ ".opam" in
-  remove_file hashes file
+  remove_file hashes file;
+  Printf.eprintf "Package %S removed\n%!" package.name
 
 let rename_package hashes package new_name =
   let new_dir = "src" // new_name in
@@ -77,39 +82,30 @@ let rename_package hashes package new_name =
       name = new_name ;
   }
 
-  (* lookup for "drom.toml" and update it *)
-let action ~package_name ~kind ~mode ~promote_skip ~dir ~create ~remove
-    ?rename
-    () =
+let action ~package_name ~kind ~mode ~promote_skip ~dir ?create ?remove
+    ?rename () =
   let p, inferred_dir = Project.get () in
   let name =
-    match package_name with None -> p.package.name | Some name -> name
+    match package_name with
+      None ->
+        let name = p.package.name in
+        Printf.eprintf "No name specified, using project name %S\n%!" name;
+        name
+    | Some name -> name
   in
-  if create then (
-    if List.exists (fun package -> package.name = name) p.packages then
-      Error.raise "A package with this name already exists";
-    let dir =
-      match dir with
-      | None ->
-          let dir = if inferred_dir = "" then "src" else inferred_dir in
-          dir // name
-      | Some dir -> dir
-    in
-    let kind = match kind with None -> Library | Some kind -> kind in
-    let package = Project.create_package ~kind ~name ~dir in
-    package.p_mode <- mode;
-    package.project <- p;
-    p.packages <- p.packages @ [ package ];
-    Update.update_files ~upgrade:true ~promote_skip:false ~git:true p )
-  else (
-    if List.for_all (fun package -> package.name <> name) p.packages then
-      Error.raise "No such package to modify";
-    if dir <> None then Error.raise "Option --dir is not available for update";
+  let upgrade =
+    Hashes.with_ctxt ~git:true (fun hashes ->
+        match remove with
+        | Some name ->
+            if create <> None then
+              Error.raise "--remove and --create are incompatible";
 
-    let upgrade =
-      Hashes.with_ctxt ~git:true (fun hashes ->
-          if remove then (
-            if p.package.name = name then Error.raise "Cannot remove main package";
+
+            if p.package.name = name then
+              Error.raise "Cannot remove main package";
+            if List.for_all (fun package -> package.name <> name)
+                p.packages then
+              Error.raise "No such package to remove";
             p.packages <-
               List.filter (fun package ->
                   if package.name = name then begin
@@ -119,43 +115,88 @@ let action ~package_name ~kind ~mode ~promote_skip ~dir ~create ~remove
                   else true
                 ) p.packages;
             true
-          )
-          else
-            match rename with
-            | Some new_name ->
-                if p.package.name = name then
-                  Error.raise "Cannot rename main package";
-                if List.exists (fun package -> package.name = new_name) p.packages
-                then
-                  Error.raise "Cannot rename to an already existing package name";
-                p.packages <-
-                  List.map (fun package ->
-                      if package.name = name then
-                        rename_package hashes package new_name
-                      else
-                        package
-                    ) p.packages;
-                true
-            | None ->
-                let upgrade = ref false in
-                List.iter
-                  (fun package ->
-                     if package.name = name then (
-                       ( match kind with
-                         | None -> ()
-                         | Some kind ->
-                             p.package.kind <- kind;
-                             upgrade := true );
-                       match mode with
-                       | None -> ()
-                       | Some mode ->
-                           p.package.p_mode <- Some mode;
-                           upgrade := true ))
-                  p.packages;
-                !upgrade
-        )
-    in
-    Update.update_files ~create:false ?mode ~upgrade ~promote_skip ~git:true p )
+        | None ->
+            let upgrade =
+              match create with
+              | Some skeleton ->
+                  if List.exists (fun package -> package.name = name)
+                      p.packages then
+                    Error.raise "A package with this name already exists";
+                  let dir =
+                    match dir with
+                    | None ->
+                        let dir = if inferred_dir = "" then "src"
+                          else inferred_dir in
+                        dir // name
+                    | Some dir -> dir
+                  in
+                  let kind = match kind with
+                      None -> Library
+                    | Some kind -> kind in
+                  let package = Project.create_package ~kind ~name ~dir in
+                  package.p_skeleton <- Some skeleton;
+                  begin match mode with
+                    | None -> ()
+                    | Some mode -> package.p_mode <- Some mode;
+                  end;
+                  package.project <- p;
+
+                  let rec iter_skeleton list =
+                    match list with
+                    | [] -> package
+                    | content :: super ->
+                        let package = iter_skeleton super in
+                        let content = Subst.package () package content in
+                        Project.package_of_string ~msg:"toml template" content
+                  in
+                  let skeleton = Skeleton.lookup_package skeleton in
+                  let package = iter_skeleton skeleton.skeleton_toml in
+                  p.packages <- p.packages @ [ package ];
+                  true
+              | None ->
+                  if List.for_all (fun package -> package.name <> name)
+                      p.packages then
+                    Error.raise "No such package to modify";
+                  if dir <> None then
+                    Error.raise "Option --dir is not available for update";
+                  match rename with
+                  | Some new_name ->
+                      if p.package.name = name then
+                        Error.raise "Cannot rename main package";
+                      if List.exists (fun package ->
+                          package.name = new_name) p.packages
+                      then
+                        Error.raise
+                          "Cannot rename to an already existing package name";
+                      p.packages <-
+                        List.map (fun package ->
+                            if package.name = name then
+                              rename_package hashes package new_name
+                            else
+                              package
+                          ) p.packages;
+                      true
+                  | None -> false
+            in
+            let upgrade = ref upgrade in
+            List.iter
+              (fun package ->
+                 if package.name = name then (
+                   ( match kind with
+                     | None -> ()
+                     | Some kind ->
+                         p.package.kind <- kind;
+                         upgrade := true );
+                   match mode with
+                   | None -> ()
+                   | Some mode ->
+                       p.package.p_mode <- Some mode;
+                       upgrade := true ))
+              p.packages;
+            !upgrade
+      )
+  in
+  Update.update_files ~create:false ?mode ~upgrade ~promote_skip ~git:true p
 
 let cmd =
   let package_name = ref None in
@@ -163,8 +204,8 @@ let cmd =
   let mode = ref None in
   let promote_skip = ref false in
   let dir = ref None in
-  let create = ref false in
-  let remove = ref false in
+  let create = ref None in
+  let remove = ref None in
   let rename = ref None in
   {
     cmd_name;
@@ -172,16 +213,16 @@ let cmd =
       (fun () ->
         action ~package_name:!package_name ~mode:!mode ~kind:!kind
           ~promote_skip:!promote_skip ~dir:!dir
-          ~create:!create ~remove:!remove
+          ?create:!create ?remove:!remove
           ?rename:!rename
           ());
     cmd_args =
       [
         ( [ "new" ],
-          Arg.Set create,
-          Ezcmd.info "Add a new package to the project" );
+          Arg.String (fun s -> create := Some s),
+          Ezcmd.info "Add a new package to the project with skeleton NAME" );
         ( [ "remove" ],
-          Arg.Set remove,
+          Arg.String (fun s -> remove := Some s),
           Ezcmd.info "Remove a package from the project" );
         ( [ "dir" ],
           Arg.String (fun s -> dir := Some s),
